Create test file for Map state handler
