+++ __tests__/StateMachine.test.ts
@@ -8,147 +8,6 @@ afterEach(() => {
 });
 
 describe('State Machine', () => {
-  describe('Map State', () => {
-    test('should iterate over the array referenced by `ItemsPath`', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'MapState',
-        States: {
-          MapState: {
-            Type: 'Map',
-            Iterator: {
-              StartAt: 'EntryIterationState',
-              States: {
-                EntryIterationState: {
-                  Type: 'Succeed',
-                },
-              },
-            },
-            ItemsPath: '$.items',
-            End: true,
-          },
-        },
-      };
-      const input = {
-        items: [
-          { num1: 5, num2: 3 },
-          { num1: 2, num2: 6 },
-          { num1: 7, num2: 4 },
-        ],
-      };
-
-      const jsonQuerySpy = jest.spyOn(StateMachine.prototype as any, 'jsonQuery');
-      const stateMachine = new StateMachine(definition);
-      const result = await stateMachine.run(input);
-
-      expect(jsonQuerySpy).toHaveBeenCalledWith('$.items', {
-        items: [
-          { num1: 5, num2: 3 },
-          { num1: 2, num2: 6 },
-          { num1: 7, num2: 4 },
-        ],
-      });
-      expect(jsonQuerySpy).toHaveReturnedWith([
-        { num1: 5, num2: 3 },
-        { num1: 2, num2: 6 },
-        { num1: 7, num2: 4 },
-      ]);
-      expect(result).toHaveLength(3);
-    });
-
-    test('should iterate over current input if `ItemsPath` is not specified', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'MapState',
-        States: {
-          MapState: {
-            Type: 'Map',
-            Iterator: {
-              StartAt: 'EntryIterationState',
-              States: {
-                EntryIterationState: {
-                  Type: 'Succeed',
-                },
-              },
-            },
-            End: true,
-          },
-        },
-      };
-      const input = [
-        { num1: 5, num2: 3 },
-        { num1: 2, num2: 6 },
-        { num1: 7, num2: 4 },
-      ];
-
-      const jsonQuerySpy = jest.spyOn(StateMachine.prototype as any, 'jsonQuery');
-      const stateMachine = new StateMachine(definition);
-      const result = await stateMachine.run(input);
-
-      expect(jsonQuerySpy).not.toHaveBeenCalled();
-      expect(result).toHaveLength(3);
-    });
-
-    test('should process `Parameters` field if specified', async () => {
-      const stateMachineDefinition: StateMachineDefinition = {
-        StartAt: 'MapState',
-        States: {
-          MapState: {
-            Type: 'Map',
-            Iterator: {
-              StartAt: 'EntryIterationState',
-              States: {
-                EntryIterationState: {
-                  Type: 'Succeed',
-                },
-              },
-            },
-            ItemsPath: '$.items',
-            Parameters: {
-              'pair.$': '$$.Map.Item.Value',
-              'index.$': '$$.Map.Item.Index',
-            },
-            End: true,
-          },
-        },
-      };
-      const input = {
-        items: [
-          { num1: 5, num2: 3 },
-          { num1: 2, num2: 6 },
-          { num1: 7, num2: 4 },
-        ],
-      };
-
-      const processPayloadTemplateSpy = jest.spyOn(StateMachine.prototype as any, 'processPayloadTemplate');
-      const stateMachine = new StateMachine(stateMachineDefinition);
-      const result = await stateMachine.run(input);
-
-      expect(processPayloadTemplateSpy).toHaveBeenCalledWith(
-        { 'pair.$': '$$.Map.Item.Value', 'index.$': '$$.Map.Item.Index' },
-        {
-          items: [
-            { num1: 5, num2: 3 },
-            { num1: 2, num2: 6 },
-            { num1: 7, num2: 4 },
-          ],
-        }
-      );
-      expect(processPayloadTemplateSpy).toHaveBeenCalledTimes(3);
-      expect(processPayloadTemplateSpy).toHaveNthReturnedWith(1, {
-        pair: { num1: 5, num2: 3 },
-        index: 0,
-      });
-      expect(processPayloadTemplateSpy).toHaveNthReturnedWith(2, {
-        pair: { num1: 2, num2: 6 },
-        index: 1,
-      });
-      expect(processPayloadTemplateSpy).toHaveNthReturnedWith(3, {
-        pair: { num1: 7, num2: 4 },
-        index: 2,
-      });
-      expect(result).toHaveLength(3);
-    });
-  });
-
   describe('Pass State', () => {
     test('should return initial input as result if neither `Result` nor `ResultPath` are specified', async () => {
       const definition: StateMachineDefinition = {
